+++ R/utilities.R
@@ -188,7 +188,7 @@ match_ind <- function(x, y, ...){
 #' @noRd
 .args_update <- function(x){
   
-  if(is.null(x)) {
+  if(!is.null(x)) {
     lapply(seq(1,length(x)), function(z){
       if(!is.null(names(x)[z])) {
         assign(names(x)[z],